+++ src/index.html
@@ -1,6 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
+  <link rel="stylesheet" href="./main.css"/>
   <link rel="stylesheet" href="./pages/Cadastro/cadastro.css"/>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -12,13 +13,9 @@
 </head>
 
 <body>  
-  <main id="root"> </main>
+  <main class="root"> </main>
 
   <script type="module" src="main.js"></script>
   <script src="/node_modules/firebase/firebase-auth.js"></script>
-
- <!--<script src="https://www.gstatic.com/firebasejs/9.9.4/firebase-app.js"></script>
- <script src="https://www.gstatic.com/firebasejs/9.9.4/firebase-firestore.js"></script> -->
-
 </body>
 </html>
