+++ src/lib/__mocks__/export.js
@@ -1,0 +1,5 @@
+export const initializeApp = jest.fn()
+export const getAuth = jest.fn()
+export const createUserWithEmailAndPassword = jest.fn()
+export const signInWithEmailAndPassword = jest.fn()
+export const updateProfile = jest.fn()
