+++ src/lib/config.js
@@ -1,4 +1,4 @@
-import { initializeApp } from 'https://www.gstatic.com/firebasejs/9.9.4/firebase-app.js';
+import { initializeApp } from "./export.js";
 //import { getFirestore } from "firebase/firestore";
 // import { initializeApp } from "firebase/app";
