+++ src/lib/export.js
@@ -1,0 +1,2 @@
+export { initializeApp } from "https://www.gstatic.com/firebasejs/9.9.4/firebase-app.js";
+export { getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword, updateProfile } from "https://www.gstatic.com/firebasejs/9.9.4/firebase-auth.js";
