+++ src/pages/Cadastro/cadastro.css
@@ -1,19 +1,31 @@
-/* estilização da página principal main */
-body {
-  text-align: center;
-  margin:0;
-  /* background: linear-gradient(to bottom right, rgb(199, 163, 232), #49AAD7, rgb(233, 173, 233), rgb(199, 163, 232)) no-repeat;
-  background-size: cover;
-  background-position: center; */
+
+
+/* estilização da página principal do container */
+.container-registration{
+  display:inline-flex;
+  align-items: center;
+  flex-direction: column;  
+  width: 100%;
   height: 100%;
 }
-body{
-  margin: 5%;
+
+/* estilização do botão voltar */
+.btn-back{
+  margin-left: 4%;
+  font-size: 20px;
+  color: white;
+  background-color: transparent;
+  border: transparent;
+  align-self: flex-start;
+  
+
 }
 
-.container-registration{
-  border:2px solid black;
-  flex-wrap: wrap;
+/* logo de coração com lampada e nome inspire */
+.logo-tittle{
+  margin-top: 0;
+  margin-bottom: 3%;
+  
 }
 
 /* estilização dos campos do nome, email e senha os campos de imput*/
@@ -21,42 +33,36 @@ body{
   display: flex;
   align-items: center;
   flex-direction: column;
+  width: 50%;
   
 }
 
-/* logo de coração com lampada */
-.one-logo{
-  align-self: center;
-}
-
-/* classe da imagem de coração com o botão voltar */
-.btnback{
-  display: flex;
-  align-items: center;
-  margin-left: 32%;
-  width: 50%;
-}
-
-/* botão voltar */
-#btn-back{
-  margin-right: 20%;
-}
+/* inputs de email, nome e senha */
 
 .input{
-  width: max-content;
-  height: max-content;
-  margin: 20px;
-  padding: 10px;
-  border-radius: 20px;
-}
+  width: 240px;
+  height: 38px;
+  border: none;
+  font-size: 14px;
+  padding: 2%;
+  border-radius: 10px;
+  box-shadow: 7px 7px 7px rgba(0, 0, 0, 0.507);
+  background-color: white;
+  margin: 2% 0;
 
+}
 
 .button{
-  padding: 10px;
-  width: 15rem;
   background-color: #49AAD7;
-  color: white;
-  font-size: medium;
-  border: 2px #49AAD7 solid;
+  color: aliceblue;
+  width: 200px;
+  height: 40px;
+  font-size: 14px;
+  margin-top: 5%;
+  font-weight:600;
+  border:none;
+  box-shadow: 5px 5px 5px rgba(0, 0, 0, 0.507);
+  margin: 2% 0;
   
 }
+
