+++ src/pages/login/login.js
@@ -1,10 +1,4 @@
-// aqui exportaras las funciones que necesites
-import { getAuth, signInWithEmailAndPassword } from "https://www.gstatic.com/firebasejs/9.9.4/firebase-auth.js";
-import { app } from "../../config-firebase.js";
-import { signIn } from "../../lib/getAuth.js";
-//export { signInWithEmailAndPassword } from "https://www.gstatic.com/firebasejs/9.9.4/firebase-auth.js";
-//export { getAuth } from "https://www.gstatic.com/firebasejs/9.9.4/firebase-auth.js";
-
+import { signIn } from "../../lib/auth.js";
 
 export default function Login() {
   const login = document.createElement('div');
@@ -42,7 +36,7 @@ export default function Login() {
   const password = login.querySelector('.password');
 
   signInButton.addEventListener("click", function (e) {
-    e.preventDefault(); 
+    e.preventDefault();
     signIn(email.value, password.value)
   })
