+++ test/lib/auth.spec.js
@@ -1,0 +1,4 @@
+import {register} from "../../src/lib/auth.js"
+import {createUserWithEmailAndPassword, updateProfile} from "../../src/lib/export.js"
+
+jest.mock("../../src/lib/export.js")
